auth-service/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-jose[cryptography]
bcrypt
python-dotenv
passlib[bcrypt]